+++ trpc/client/mysql/BUILD
@@ -91,22 +91,6 @@ filegroup(
 )
 
 cc_test(
-    name = "mysql_executor_pool_test",
-    srcs = ["mysql_executor_pool_test.cc"],
-    data = [
-        ":test_yaml_files"
-    ],
-    visibility = ["//visibility:public"],
-    deps = [
-        ":mysql_executor",
-        ":mysql_executor_pool",
-        "//trpc/common/config:trpc_config",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-cc_test(
     name = "mysql_executor_pool_manager_test",
     srcs = ["mysql_executor_pool_manager_test.cc"],
     data = [
