+++ trpc/client/mysql/mysql_executor_pool_manager_test.cc
@@ -3,6 +3,7 @@
 #include <iostream>
 #include <thread>
 #include "trpc/client/mysql/mysql_executor.h"
+#include "trpc/client/mysql/mysql_executor_pool.h"
 #include "trpc/common/config/mysql_client_conf.h"
 #include "trpc/common/config/trpc_config.h"
 
@@ -40,9 +41,8 @@ void op1(int begin, int end) {
   }
 }
 
-void op2(MysqlExecutorPoolManager* manager, const MysqlClientConf& conf, int begin, int end) {
+void op2(MysqlExecutorPool* pool, int begin, int end) {
   for (int i = begin; i < end; i++) {
-    auto pool = manager->getPool(conf);  // Get the pool with conf parameter
     shared_ptr<MysqlExecutor> conn = pool->getConnection();
     char sql[1024] = {0};
     sprintf(sql, "insert into person values(%d, 18, 'man', 'starry')", i);
@@ -64,9 +64,10 @@ TEST(PerformanceTest, SingleThreadedWithoutPool) {
 TEST(PerformanceTest, SingleThreadedWithPool) {
   clearPersonTable();  // 清空 person 表数据
   MysqlClientConf conf = readMysqlClientConf();
-  MysqlExecutorPoolManager manager;  // 使用构造函数创建 manager 实例
+  // MysqlExecutorPool* pool = MysqlExecutorPoolManager::getPool(conf);
+  MysqlExecutorPool* pool = new MysqlExecutorPool(conf);
   steady_clock::time_point begin = steady_clock::now();
-  op2(&manager, conf, 0, 500);  // 传递 manager 和 conf
+  op2(pool, 0, 500);
   steady_clock::time_point end = steady_clock::now();
   auto length = end - begin;
   cout << "连接池，单线程，用时：" << length.count() << "纳秒，" << length.count() / 1000000 << "毫秒" << endl;
@@ -93,13 +94,13 @@ TEST(PerformanceTest, MultiThreadedWithoutPool) {
 TEST(PerformanceTest, MultiThreadedWithPool) {
   clearPersonTable();  // 清空 person 表数据
   MysqlClientConf conf = readMysqlClientConf();
-  MysqlExecutorPoolManager manager;  // 使用构造函数创建 manager 实例
+  MysqlExecutorPool* pool = MysqlExecutorPoolManager::getPool(conf);
   steady_clock::time_point begin = steady_clock::now();
-  thread t1(op2, &manager, conf, 0, 100);
-  thread t2(op2, &manager, conf, 100, 200);
-  thread t3(op2, &manager, conf, 200, 300);
-  thread t4(op2, &manager, conf, 300, 400);
-  thread t5(op2, &manager, conf, 400, 500);
+  thread t1(op2, pool, 0, 100);
+  thread t2(op2, pool, 100, 200);
+  thread t3(op2, pool, 200, 300);
+  thread t4(op2, pool, 300, 400);
+  thread t5(op2, pool, 400, 500);
   t1.join();
   t2.join();
   t3.join();
